Move executor into default folder
