+++ executors/default/go.mod
@@ -1,4 +1,4 @@
-module github.com/Azure/orkestra-workflow-executor
+module github.com/Azure/orkestra-workflow-executor/executors/default
 
 go 1.16
 
@@ -10,7 +10,6 @@ require (
 	k8s.io/apimachinery v0.21.0
 	k8s.io/client-go v0.21.0
 	k8s.io/kubectl v0.20.4
-	sigs.k8s.io/cli-utils v0.25.0
 	sigs.k8s.io/controller-runtime v0.8.3
 	sigs.k8s.io/yaml v1.2.0
 )
